requirements.txt
# Azure Functions
azure-functions>=1.18.0

# Azure AI Services
azure-search-documents>=11.4.0
azure-ai-inference>=1.0.0b5
azure-identity>=1.15.0
openai>=1.12.0

# Azure Storage (for conversation persistence)
azure-data-tables>=12.4.0

# Application Insights
opencensus-ext-azure>=1.1.13
opencensus-ext-logging>=0.1.1

# Core dependencies
python-dotenv>=1.0.0
tiktoken>=0.5.2
requests>=2.31.0
